Move life cycles to 'core' package
