+++ src/main/java/ru/shishmakov/core/ServiceController.java
@@ -3,7 +3,6 @@ package ru.shishmakov.core;
 import com.google.common.util.concurrent.MoreExecutors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import ru.shishmakov.util.LifeCycle;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
@@ -17,7 +16,7 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.SECONDS;
-import static ru.shishmakov.util.LifeCycle.*;
+import static ru.shishmakov.core.LifeCycle.*;
 import static ru.shishmakov.util.Threads.*;
 
 /**
